requirements.txt
asgiref==3.3.1
certifi==2020.12.5
Django==1.11.8
django-filter==1.1.0
djangorestframework==3.7.3
importlib-metadata==3.4.0
Markdown==3.3.3
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
